clean test utils
